+++ examples/static-hbase/main.tf
@@ -1,3 +1,7 @@
+locals {
+  this_application = ["Hbase"]
+}
+
 # Set up logs bucket with read/write permissions
 module "emr-logs-bucket" {
   source      = "git::git@github.com:Datatamer/terraform-aws-s3.git?ref=1.1.0"
@@ -47,13 +51,13 @@ resource "aws_s3_bucket_object" "sample_bootstrap_script_2" {
 
 # EMR Static HBase cluster
 module "emr-hbase" {
-  # source = "git::git@github.com:Datatamer/terraform-aws-emr.git?ref=5.0.0"
+  # source = "git::git@github.com:Datatamer/terraform-aws-emr.git?ref=6.0.0"
   source = "../.."
 
   # Configurations
   create_static_cluster         = true
   release_label                 = "emr-5.29.0" # hbase 1.4.10
-  applications                  = ["Hbase"]
+  applications                  = local.this_application
   emr_config_file_path          = "../emr-config-template.json"
   bucket_path_to_logs           = "logs/hbase-test-cluster/"
   json_configuration_bucket_key = "tamr/emr/emr.json"
@@ -73,10 +77,8 @@ module "emr-hbase" {
   ]
 
   # Networking
-  subnet_id  = var.subnet_id
-  vpc_id     = var.vpc_id
-  tamr_cidrs = []
-  tamr_sgs   = []
+  subnet_id = var.subnet_id
+  vpc_id    = var.vpc_id
 
   # External resource references
   bucket_name_for_root_directory = module.emr-rootdir-bucket.bucket_name
@@ -85,19 +87,17 @@ module "emr-hbase" {
   key_pair_name                  = module.emr_key_pair.key_pair_key_name
 
   # Names
-  cluster_name                  = "HBase-Test-EMR-Cluster"
-  emr_service_role_name         = "hbase-test-service-role"
-  emr_ec2_role_name             = "hbase-test-ec2-role"
-  emr_ec2_instance_profile_name = "hbase-test-instance-profile"
-  emr_service_iam_policy_name   = "hbase-test-service-policy"
-  emr_ec2_iam_policy_name       = "hbase-test-ec2-policy"
-  master_instance_fleet_name    = "HBase-Test-MasterInstanceFleet"
-  core_instance_fleet_name      = "Hbase-Test-CoreInstanceFleet"
-  emr_managed_master_sg_name    = "HBase-Test-EMR-Hbase-Master"
-  emr_managed_core_sg_name      = "HBase-Test-EMR-Hbase-Core"
-  emr_additional_master_sg_name = "HBase-Test-EMR-Hbase-Additional-Master"
-  emr_additional_core_sg_name   = "HBase-Test-EMR-Hbase-Additional-Core"
-  emr_service_access_sg_name    = "HBase-Test-EMR-Hbase-Service-Access"
+  cluster_name                  = format("%s-%s", var.name_prefix, "HBase-Test-EMR-Cluster")
+  emr_service_role_name         = format("%s-%s", var.name_prefix, "hbase-test-service-role")
+  emr_ec2_role_name             = format("%s-%s", var.name_prefix, "hbase-test-ec2-role")
+  emr_ec2_instance_profile_name = format("%s-%s", var.name_prefix, "hbase-test-instance-profile")
+  emr_service_iam_policy_name   = format("%s-%s", var.name_prefix, "hbase-test-service-policy")
+  emr_ec2_iam_policy_name       = format("%s-%s", var.name_prefix, "hbase-test-ec2-policy")
+  master_instance_fleet_name    = format("%s-%s", var.name_prefix, "HBase-Test-MasterInstanceFleet")
+  core_instance_fleet_name      = format("%s-%s", var.name_prefix, "Hbase-Test-CoreInstanceFleet")
+  emr_managed_master_sg_name    = format("%s-%s", var.name_prefix, "HBase-Test-EMR-Hbase-Master")
+  emr_managed_core_sg_name      = format("%s-%s", var.name_prefix, "HBase-Test-EMR-Hbase-Core")
+  emr_service_access_sg_name    = format("%s-%s", var.name_prefix, "HBase-Test-EMR-Hbase-Service-Access")
 
   # Scale
   master_instance_on_demand_count = 1
@@ -107,10 +107,50 @@ module "emr-hbase" {
   master_ebs_size                 = 50
   core_ebs_size                   = 50
 
-  # Spot Instance definition
-  # on-demand r5.xlarge EC2 + EMR --> $0.252 + $0.063 = $0.315 per Hour
-  core_bid_price = ".100"
-
   # Don't use spot instances for the master nodes
   master_bid_price = ""
+
+  # Security Group IDs
+  emr_managed_master_sg_ids = module.aws-emr-sg-master.security_group_ids
+  emr_managed_core_sg_ids   = module.aws-emr-sg-core.security_group_ids
+  emr_service_access_sg_ids = module.aws-emr-sg-service-access.security_group_ids
+}
+
+module "sg-ports" {
+  # source               = "git::https://github.com/Datatamer/terraform-aws-emr.git//modules/aws-emr-ports?ref=6.0.0"
+  source       = "../../modules/aws-emr-ports"
+  applications = local.this_application
+}
+
+module "aws-emr-sg-master" {
+  source              = "git::git@github.com:Datatamer/terraform-aws-security-groups.git?ref=1.0.0"
+  vpc_id              = var.vpc_id
+  ingress_cidr_blocks = var.ingress_cidr_blocks
+  egress_cidr_blocks  = var.egress_cidr_blocks
+  ingress_ports       = module.sg-ports.ingress_master_ports
+  sg_name_prefix      = format("%s-%s", var.name_prefix, "-master")
+  egress_protocol     = "all"
+  ingress_protocol    = "tcp"
+}
+
+module "aws-emr-sg-core" {
+  source              = "git::git@github.com:Datatamer/terraform-aws-security-groups.git?ref=1.0.0"
+  vpc_id              = var.vpc_id
+  ingress_cidr_blocks = var.ingress_cidr_blocks
+  egress_cidr_blocks  = var.egress_cidr_blocks
+  ingress_ports       = module.sg-ports.ingress_core_ports
+  sg_name_prefix      = format("%s-%s", var.name_prefix, "-core")
+  egress_protocol     = "all"
+  ingress_protocol    = "tcp"
+}
+
+module "aws-emr-sg-service-access" {
+  source              = "git::git@github.com:Datatamer/terraform-aws-security-groups.git?ref=1.0.0"
+  vpc_id              = var.vpc_id
+  ingress_cidr_blocks = var.ingress_cidr_blocks
+  egress_cidr_blocks  = var.egress_cidr_blocks
+  ingress_ports       = module.sg-ports.ingress_service_access_ports
+  sg_name_prefix      = format("%s-%s", var.name_prefix, "-service-access")
+  egress_protocol     = "all"
+  ingress_protocol    = "tcp"
 }