+++ examples/static-spark/variables.tf
@@ -18,6 +18,22 @@ variable "bucket_name_for_logs" {
   description = "S3 bucket name for cluster logs."
 }
 
+variable "name_prefix" {
+  type        = string
+  description = "A string to prepend to names of the resources in the cluster"
+}
+
+variable "ingress_cidr_blocks" {
+  description = "CIDR blocks to attach to security groups for ingress"
+  type        = list(string)
+}
+
+variable "egress_cidr_blocks" {
+  description = "CIDR blocks to attach to security groups for egress"
+  type        = list(string)
+  default     = ["0.0.0.0/0"]
+}
+
 variable "tags" {
   type        = map(string)
   description = "A map of tags to add to all resources created by this example."
@@ -26,5 +42,3 @@ variable "tags" {
     Environment = "Example"
   }
 }
-
-